cloud_function.py
import os
from google.cloud import storage, vision

def count_objects(data, context):
    # Extract the bucket and file information from the event data
    bucket_name = data['bucket']
    file_name = data['name']

    # Only process photos and videos with the specified prefixes
    if not file_name.startswith(('photo', 'video')):
        return

    # Set up the Cloud Vision API client
    client = vision.ImageAnnotatorClient()

    # Get the updated file path for saving annotated file
    updated_file_path = os.path.join('/tmp', file_name)

    # Download the file from Cloud Storage to a temporary location
    storage_client = storage.Client()
    bucket = storage_client.get_bucket(bucket_name)
    blob = bucket.blob(file_name)
    blob.download_to_filename(updated_file_path)

    # Perform object detection based on file type
    if file_name.startswith('photo'):
        # Read the image file
        with open(updated_file_path, 'rb') as image_file:
            content = image_file.read()

        # Create an image instance
        image = vision.Image(content=content)

        # Perform object detection
        response = client.object_localization(image=image)
        objects = response.localized_object_annotations

        # Add object count annotation to the image
        image_with_objects = vision.ImageAnnotatorClient.annotate_image(
            image=image,
            features=[vision.Feature(type_=vision.Feature.Type.OBJECT_LOCALIZATION)],
        )

        # Save the updated image with object count annotation
        image_with_objects.save(updated_file_path)

    elif file_name.startswith('video'):
        # Perform object detection on each frame of the video
        # Implementation of video object detection depends on the specific requirements and libraries used.
        # You may need to use tools like OpenCV or a dedicated video processing library to process the frames.

        # Sample implementation:
        # Read the video file using OpenCV and process each frame
        import cv2

        # Set up the video capture
        video = cv2.VideoCapture(updated_file_path)

        # Create a writer to save the updated video with object count annotation
        frame_width = int(video.get(cv2.CAP_PROP_FRAME_WIDTH))
        frame_height = int(video.get(cv2.CAP_PROP_FRAME_HEIGHT))
        fps = video.get(cv2.CAP_PROP_FPS)
        writer = cv2.VideoWriter(updated_file_path, cv2.VideoWriter_fourcc(*'mp4v'), fps, (frame_width, frame_height))

        # Process each frame and perform object detection
        while video.isOpened():
            ret, frame = video.read()
            if not ret:
                break

            # Convert the frame to an image instance
            image = vision.Image(content=cv2.imencode('.jpg', frame)[1].tostring())

            # Perform object detection
            response = client.object_localization(image=image)
            objects = response.localized_object_annotations

            # Draw object count annotation on the frame
            # Modify this section to suit your specific requirements for drawing object count on each frame

            # Write the frame with object count annotation to the video writer
            writer.write(frame)

        # Release the video capture and writer
        video.release()
        writer.release()

    else:
        return

    # Upload the updated file back to Cloud Storage
    updated_blob = bucket.blob(file_name)
    updated_blob.upload_from_filename(updated_file_path)

    # Delete the temporary file
    os.remove(updated_file_path)

    # Log the object count and updated file path
    print(f"Object count: {len(objects)}")
    print(f"Updated file path: gs://{bucket_name}/{file_name}")

    # Return the updated file path
    return f"gs://{bucket_name}/{file_name}"
