requirements.txt
google-cloud-storage
google-cloud-vision
opencv-python-headless
numpy
